requirements.txt
pyyaml
jinja2
typer
rich
pydantic
jsonschema
